back-end/requirements.txt
django>=2.1.11
graphene-django==2.2.0
django-filter==2.0.0
django-graphql-jwt==0.1.5
gunicorn
dj-database-url
